+++ test-packages/test-services/v2/test-service/Casetest_1Api.d.ts
@@ -6,8 +6,7 @@ import {
   DeSerializers,
   AllFields,
   EntityBuilderType,
-  EntityApi,
-  EdmTypeField
+  EntityApi
 } from '@sap-cloud-sdk/odata-v2';
 export declare class Casetest_1Api<
   DeSerializersT extends DeSerializers = DefaultDeSerializers
@@ -59,7 +58,7 @@ export declare class Casetest_1Api<
      * Static representation of the [[keyPropertyString]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    KEY_PROPERTY_STRING: EdmTypeField<
+    KEY_PROPERTY_STRING: import('@sap-cloud-sdk/odata-v2').EdmTypeField<
       Casetest_1<
         DeSerializers<
           any,