+++ test-packages/test-services/v2/test-service/TestEntityLvl2SingleLinkApi.d.ts
@@ -6,9 +6,7 @@ import {
   DeSerializers,
   AllFields,
   EntityBuilderType,
-  EntityApi,
-  EdmTypeField,
-  OrderableEdmTypeField
+  EntityApi
 } from '@sap-cloud-sdk/odata-v2';
 export declare class TestEntityLvl2SingleLinkApi<
   DeSerializersT extends DeSerializers = DefaultDeSerializers
@@ -65,7 +63,7 @@ export declare class TestEntityLvl2SingleLinkApi<
      * Static representation of the [[keyProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    KEY_PROPERTY: EdmTypeField<
+    KEY_PROPERTY: import('@sap-cloud-sdk/odata-v2').EdmTypeField<
       TestEntityLvl2SingleLink<
         DeSerializers<
           any,
@@ -96,7 +94,7 @@ export declare class TestEntityLvl2SingleLinkApi<
      * Static representation of the [[stringProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    STRING_PROPERTY: EdmTypeField<
+    STRING_PROPERTY: import('@sap-cloud-sdk/odata-v2').EdmTypeField<
       TestEntityLvl2SingleLink<
         DeSerializers<
           any,
@@ -127,7 +125,7 @@ export declare class TestEntityLvl2SingleLinkApi<
      * Static representation of the [[booleanProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    BOOLEAN_PROPERTY: EdmTypeField<
+    BOOLEAN_PROPERTY: import('@sap-cloud-sdk/odata-v2').EdmTypeField<
       TestEntityLvl2SingleLink<
         DeSerializers<
           any,
@@ -158,7 +156,7 @@ export declare class TestEntityLvl2SingleLinkApi<
      * Static representation of the [[guidProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    GUID_PROPERTY: EdmTypeField<
+    GUID_PROPERTY: import('@sap-cloud-sdk/odata-v2').EdmTypeField<
       TestEntityLvl2SingleLink<
         DeSerializers<
           any,
@@ -189,7 +187,7 @@ export declare class TestEntityLvl2SingleLinkApi<
      * Static representation of the [[int16Property]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    INT_16_PROPERTY: OrderableEdmTypeField<
+    INT_16_PROPERTY: import('@sap-cloud-sdk/odata-v2').OrderableEdmTypeField<
       TestEntityLvl2SingleLink<
         DeSerializers<
           any,