+++ test-packages/test-services/v2/test-service/TestEntityWithSharedEntityType2Api.d.ts
@@ -6,8 +6,7 @@ import {
   DeSerializers,
   AllFields,
   EntityBuilderType,
-  EntityApi,
-  EdmTypeField
+  EntityApi
 } from '@sap-cloud-sdk/odata-v2';
 export declare class TestEntityWithSharedEntityType2Api<
   DeSerializersT extends DeSerializers = DefaultDeSerializers
@@ -64,7 +63,7 @@ export declare class TestEntityWithSharedEntityType2Api<
      * Static representation of the [[keyProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    KEY_PROPERTY: EdmTypeField<
+    KEY_PROPERTY: import('@sap-cloud-sdk/odata-v2').EdmTypeField<
       TestEntityWithSharedEntityType2<
         DeSerializers<
           any,