+++ test-packages/test-services/v4/multiple-schemas-service/TestEntity1Api.d.ts
@@ -8,10 +8,7 @@ import {
   DeSerializers,
   AllFields,
   EntityBuilderType,
-  EntityApi,
-  EdmTypeField,
-  OrderableEdmTypeField,
-  EnumField
+  EntityApi
 } from '@sap-cloud-sdk/odata-v4';
 export declare class TestEntity1Api<
   DeSerializersT extends DeSerializers = DefaultDeSerializers
@@ -65,7 +62,7 @@ export declare class TestEntity1Api<
      * Static representation of the [[keyPropertyString]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    KEY_PROPERTY_STRING: EdmTypeField<
+    KEY_PROPERTY_STRING: import('@sap-cloud-sdk/odata-v4').EdmTypeField<
       TestEntity1<
         DeSerializers<
           any,
@@ -98,7 +95,7 @@ export declare class TestEntity1Api<
      * Static representation of the [[int16Property]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    INT_16_PROPERTY: OrderableEdmTypeField<
+    INT_16_PROPERTY: import('@sap-cloud-sdk/odata-v4').OrderableEdmTypeField<
       TestEntity1<
         DeSerializers<
           any,
@@ -131,7 +128,7 @@ export declare class TestEntity1Api<
      * Static representation of the [[enumProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    ENUM_PROPERTY: EnumField<
+    ENUM_PROPERTY: import('@sap-cloud-sdk/odata-v4').EnumField<
       TestEntity1<
         DeSerializers<
           any,