+++ test-packages/test-services/v4/test-service/TestEntityApi.d.ts
@@ -13,10 +13,6 @@ import {
   AllFields,
   EntityBuilderType,
   EntityApi,
-  EdmTypeField,
-  OrderableEdmTypeField,
-  CollectionField,
-  EnumField,
   OneToManyLink,
   OneToOneLink
 } from '@sap-cloud-sdk/odata-v4';
@@ -105,7 +101,7 @@ export declare class TestEntityApi<
      * Static representation of the [[keyPropertyGuid]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    KEY_PROPERTY_GUID: EdmTypeField<
+    KEY_PROPERTY_GUID: import('@sap-cloud-sdk/odata-v4').EdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -138,7 +134,7 @@ export declare class TestEntityApi<
      * Static representation of the [[keyPropertyString]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    KEY_PROPERTY_STRING: EdmTypeField<
+    KEY_PROPERTY_STRING: import('@sap-cloud-sdk/odata-v4').EdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -171,7 +167,7 @@ export declare class TestEntityApi<
      * Static representation of the [[stringProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    STRING_PROPERTY: EdmTypeField<
+    STRING_PROPERTY: import('@sap-cloud-sdk/odata-v4').EdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -204,7 +200,7 @@ export declare class TestEntityApi<
      * Static representation of the [[booleanProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    BOOLEAN_PROPERTY: EdmTypeField<
+    BOOLEAN_PROPERTY: import('@sap-cloud-sdk/odata-v4').EdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -237,7 +233,7 @@ export declare class TestEntityApi<
      * Static representation of the [[guidProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    GUID_PROPERTY: EdmTypeField<
+    GUID_PROPERTY: import('@sap-cloud-sdk/odata-v4').EdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -270,7 +266,7 @@ export declare class TestEntityApi<
      * Static representation of the [[int16Property]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    INT_16_PROPERTY: OrderableEdmTypeField<
+    INT_16_PROPERTY: import('@sap-cloud-sdk/odata-v4').OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -303,7 +299,7 @@ export declare class TestEntityApi<
      * Static representation of the [[int32Property]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    INT_32_PROPERTY: OrderableEdmTypeField<
+    INT_32_PROPERTY: import('@sap-cloud-sdk/odata-v4').OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -336,7 +332,7 @@ export declare class TestEntityApi<
      * Static representation of the [[int64Property]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    INT_64_PROPERTY: OrderableEdmTypeField<
+    INT_64_PROPERTY: import('@sap-cloud-sdk/odata-v4').OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -369,7 +365,7 @@ export declare class TestEntityApi<
      * Static representation of the [[decimalProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    DECIMAL_PROPERTY: OrderableEdmTypeField<
+    DECIMAL_PROPERTY: import('@sap-cloud-sdk/odata-v4').OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -402,7 +398,7 @@ export declare class TestEntityApi<
      * Static representation of the [[singleProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    SINGLE_PROPERTY: OrderableEdmTypeField<
+    SINGLE_PROPERTY: import('@sap-cloud-sdk/odata-v4').OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -435,7 +431,7 @@ export declare class TestEntityApi<
      * Static representation of the [[doubleProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    DOUBLE_PROPERTY: OrderableEdmTypeField<
+    DOUBLE_PROPERTY: import('@sap-cloud-sdk/odata-v4').OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -468,7 +464,7 @@ export declare class TestEntityApi<
      * Static representation of the [[floatProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    FLOAT_PROPERTY: OrderableEdmTypeField<
+    FLOAT_PROPERTY: import('@sap-cloud-sdk/odata-v4').OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -501,7 +497,7 @@ export declare class TestEntityApi<
      * Static representation of the [[timeOfDayProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    TIME_OF_DAY_PROPERTY: OrderableEdmTypeField<
+    TIME_OF_DAY_PROPERTY: import('@sap-cloud-sdk/odata-v4').OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -534,7 +530,7 @@ export declare class TestEntityApi<
      * Static representation of the [[dateProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    DATE_PROPERTY: OrderableEdmTypeField<
+    DATE_PROPERTY: import('@sap-cloud-sdk/odata-v4').OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -567,7 +563,7 @@ export declare class TestEntityApi<
      * Static representation of the [[dateTimeOffSetProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    DATE_TIME_OFF_SET_PROPERTY: OrderableEdmTypeField<
+    DATE_TIME_OFF_SET_PROPERTY: import('@sap-cloud-sdk/odata-v4').OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -600,7 +596,7 @@ export declare class TestEntityApi<
      * Static representation of the [[durationProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    DURATION_PROPERTY: OrderableEdmTypeField<
+    DURATION_PROPERTY: import('@sap-cloud-sdk/odata-v4').OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -633,7 +629,7 @@ export declare class TestEntityApi<
      * Static representation of the [[byteProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    BYTE_PROPERTY: OrderableEdmTypeField<
+    BYTE_PROPERTY: import('@sap-cloud-sdk/odata-v4').OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -666,7 +662,7 @@ export declare class TestEntityApi<
      * Static representation of the [[sByteProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    S_BYTE_PROPERTY: OrderableEdmTypeField<
+    S_BYTE_PROPERTY: import('@sap-cloud-sdk/odata-v4').OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -699,7 +695,7 @@ export declare class TestEntityApi<
      * Static representation of the [[geographyPointProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    GEOGRAPHY_POINT_PROPERTY: EdmTypeField<
+    GEOGRAPHY_POINT_PROPERTY: import('@sap-cloud-sdk/odata-v4').EdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -732,7 +728,7 @@ export declare class TestEntityApi<
      * Static representation of the [[somethingTheSdkDoesNotSupport]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    SOMETHING_THE_SDK_DOES_NOT_SUPPORT: EdmTypeField<
+    SOMETHING_THE_SDK_DOES_NOT_SUPPORT: import('@sap-cloud-sdk/odata-v4').EdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -765,7 +761,7 @@ export declare class TestEntityApi<
      * Static representation of the [[collectionProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    COLLECTION_PROPERTY: CollectionField<
+    COLLECTION_PROPERTY: import('@sap-cloud-sdk/odata-v4').CollectionField<
       TestEntity<
         DeSerializers<
           any,
@@ -830,7 +826,7 @@ export declare class TestEntityApi<
      * Static representation of the [[complexTypeCollectionProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    COMPLEX_TYPE_COLLECTION_PROPERTY: CollectionField<
+    COMPLEX_TYPE_COLLECTION_PROPERTY: import('@sap-cloud-sdk/odata-v4').CollectionField<
       TestEntity<
         DeSerializers<
           any,
@@ -863,7 +859,7 @@ export declare class TestEntityApi<
      * Static representation of the [[enumProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    ENUM_PROPERTY: EnumField<
+    ENUM_PROPERTY: import('@sap-cloud-sdk/odata-v4').EnumField<
       TestEntity<
         DeSerializers<
           any,
@@ -896,7 +892,7 @@ export declare class TestEntityApi<
      * Static representation of the [[enumPropertyInt64]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    ENUM_PROPERTY_INT_64: EnumField<
+    ENUM_PROPERTY_INT_64: import('@sap-cloud-sdk/odata-v4').EnumField<
       TestEntity<
         DeSerializers<
           any,
@@ -929,7 +925,7 @@ export declare class TestEntityApi<
      * Static representation of the [[enumPropertyWithOneMember]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    ENUM_PROPERTY_WITH_ONE_MEMBER: EnumField<
+    ENUM_PROPERTY_WITH_ONE_MEMBER: import('@sap-cloud-sdk/odata-v4').EnumField<
       TestEntity<
         DeSerializers<
           any,
@@ -962,7 +958,7 @@ export declare class TestEntityApi<
      * Static representation of the [[enumCollectionProperty]] property for query construction.
      * Use to reference this property in query operations such as 'select' in the fluent request API.
      */
-    ENUM_COLLECTION_PROPERTY: CollectionField<
+    ENUM_COLLECTION_PROPERTY: import('@sap-cloud-sdk/odata-v4').CollectionField<
       TestEntity<
         DeSerializers<
           any,