+++ CHANGELOG.md
@@ -27,7 +27,7 @@
 
 ## Improvements
 
-- [generator] Remove unused imports (`moment` and `bignumber.js`) in API classes (e.g., `BusinessPartnerAPI`) to reduce memory usage.
+- [generator] Remove unused imports (e.g, `moment`, `bignumber.js` and `EdmTypeField`) in API classes (e.g., `BusinessPartnerAPI`) to reduce memory usage.
 - [http-client] Introduce consistent query parameter encoding for all non custom parameters.
 
 ## Fixed Issues